refactor: align domain files with folder responsibilities
