+++ dogArea/Source/Domain/Home/Services/HomeWeeklyStatisticsService.swift
