+++ dogArea/Source/Domain/Map/Models/MapModel.swift
