+++ dogArea/Views/MapView/MapView.swift
@@ -663,89 +663,3 @@ private extension MapTopBannerKind {
         }
     }
 }
-
-enum RecoveryIssueKind: Equatable {
-    case locationPermissionDenied
-    case networkOffline
-    case authExpired
-}
-
-struct RecoveryIssue: Identifiable, Equatable {
-    let kind: RecoveryIssueKind
-    let detail: String?
-
-    var id: String {
-        "\(kind)-\(detail ?? "")"
-    }
-
-    var title: String {
-        switch kind {
-        case .locationPermissionDenied:
-            return "위치 권한이 필요해요"
-        case .networkOffline:
-            return "오프라인 모드"
-        case .authExpired:
-            return "인증이 만료됐어요"
-        }
-    }
-
-    var message: String {
-        switch kind {
-        case .locationPermissionDenied:
-            return "설정에서 위치 권한을 허용하면 산책 기록을 계속할 수 있어요."
-        case .networkOffline:
-            return "지금 기록은 기기에 저장되고, 온라인 복귀 시 자동 동기화돼요."
-        case .authExpired:
-            return "다시 로그인하면 현재 화면으로 돌아와서 이어서 진행할 수 있어요."
-        }
-    }
-
-    var primaryButtonTitle: String {
-        switch kind {
-        case .locationPermissionDenied:
-            return "설정 열기"
-        case .networkOffline:
-            return "다시 시도"
-        case .authExpired:
-            return "다시 로그인"
-        }
-    }
-}
-
-enum RecoveryIssueClassifier {
-    static func fromSyncErrorCode(_ rawValue: String?) -> RecoveryIssue? {
-        guard let rawValue, rawValue.isEmpty == false else { return nil }
-        switch rawValue {
-        case SyncOutboxErrorCode.offline.rawValue:
-            return RecoveryIssue(kind: .networkOffline, detail: rawValue)
-        case SyncOutboxErrorCode.tokenExpired.rawValue, SyncOutboxErrorCode.unauthorized.rawValue:
-            return RecoveryIssue(kind: .authExpired, detail: rawValue)
-        default:
-            return nil
-        }
-    }
-
-    static func fromErrorMessage(_ raw: String?) -> RecoveryIssue? {
-        guard let raw else { return nil }
-        let normalized = raw.lowercased()
-        if normalized.contains("network"),
-           normalized.contains("offline") || normalized.contains("internet") {
-            return RecoveryIssue(kind: .networkOffline, detail: raw)
-        }
-        if normalized.contains("token") || normalized.contains("unauthorized") || normalized.contains("auth") {
-            return RecoveryIssue(kind: .authExpired, detail: raw)
-        }
-        return nil
-    }
-}
-
-enum RecoverySystemAction {
-    static func openAppSettings() {
-        #if canImport(UIKit)
-        guard let url = URL(string: UIApplication.openSettingsURLString) else { return }
-        if UIApplication.shared.canOpenURL(url) {
-            UIApplication.shared.open(url)
-        }
-        #endif
-    }
-}
